Reuse whisper's pcm_decode
